pyproject.toml
[tool.poetry]
name = "shavatar"
version = "1.1.0"
description = "SHA-based avatar generation."
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
